refactor(frontend): move Tag to frontend/
